split webpack and dev server files
