+++ src/client/App.js
@@ -2,7 +2,6 @@ import React, { useContext } from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Navbar from './components/Navbar';
 import Home from './pages/Home';
-import Test from './pages/Test';
 
 import { SidebarContext } from './context/sidebarContext';
